# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


export const QuestionType = {
  text: 'text',
  choice: 'choice',
  multiChoice: 'multiChoice',
} as const;

export interface Question {
  id: string;
  question: string;
  type: QuestionType;
  options?: string[];
  placeholder?: string;
}

export interface UploadResult {
  sessionId: string;
  extractedContent: string;
  questions: Question[];
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  uploading: 'uploading',
  questionnaire: 'questionnaire',
  generating: 'generating',
  complete: 'complete',
} as const;

/**
 * @nullable
 */
export type SessionAnswers = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type SessionNotes = { [key: string]: unknown } | null;

export interface Session {
  sessionId: string;
  status: SessionStatus;
  /** @nullable */
  extractedContent?: string | null;
  questions: Question[];
  /** @nullable */
  answers?: SessionAnswers;
  /** @nullable */
  notes?: SessionNotes;
  uploadedImageUrls?: string[];
}

export type QuestionnaireInputAnswers = { [key: string]: unknown };

export interface QuestionnaireInput {
  answers: QuestionnaireInputAnswers;
  /** @nullable */
  additionalInstructions?: string | null;
}

export interface QuestionnaireResult {
  needsMoreInfo: boolean;
  followUpQuestions: Question[];
  ready: boolean;
  /** @nullable */
  message?: string | null;
}

export interface EditInput {
  instructions: string;
  uploadedImageUrls?: string[];
}

export type NotePageType = typeof NotePageType[keyof typeof NotePageType];


export const NotePageType = {
  cover: 'cover',
  section: 'section',
  content: 'content',
  qa: 'qa',
  reflection: 'reflection',
} as const;

export interface NoteItem {
  t: string;
  /** @nullable */
  n?: string | null;
}

export interface NoteSubsection {
  title: string;
  items?: NoteItem[];
}

export interface NoteSection {
  title: string;
  items?: NoteItem[];
  subsections?: NoteSubsection[];
}

export interface ReflectionQuestion {
  q: string;
}

export interface NotePage {
  id: string;
  type: NotePageType;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  band?: string | null;
  /** @nullable */
  eventName?: string | null;
  /** @nullable */
  eventDate?: string | null;
  sections?: NoteSection[];
  /** @nullable */
  quote?: string | null;
  /** @nullable */
  quoteSource?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  questions?: ReflectionQuestion[];
}

export interface GeneratedNotes {
  sessionId: string;
  pages: NotePage[];
  uploadedImageUrls?: string[];
}

export interface ErrorResponse {
  error: string;
}