# lib/api-zod/src/generated/types/generatedNotes.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotePage } from './notePage';

export interface GeneratedNotes {
  sessionId: string;
  pages: NotePage[];
  uploadedImageUrls?: string[];
}