# lib/api-zod/src/generated/types/notePage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotePageType } from './notePageType';
import type { NoteSection } from './noteSection';
import type { ReflectionQuestion } from './reflectionQuestion';

export interface NotePage {
  id: string;
  type: NotePageType;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  band?: string | null;
  /** @nullable */
  eventName?: string | null;
  /** @nullable */
  eventDate?: string | null;
  sections?: NoteSection[];
  /** @nullable */
  quote?: string | null;
  /** @nullable */
  quoteSource?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  questions?: ReflectionQuestion[];
}