# lib/api-zod/src/generated/types/noteSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NoteItem } from './noteItem';
import type { NoteSubsection } from './noteSubsection';

export interface NoteSection {
  title: string;
  items?: NoteItem[];
  subsections?: NoteSubsection[];
}