# lib/api-zod/src/generated/types/noteSubsection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NoteItem } from './noteItem';

export interface NoteSubsection {
  title: string;
  items?: NoteItem[];
}