# lib/api-zod/src/generated/types/questionnaireInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionnaireInputAnswers } from './questionnaireInputAnswers';

export interface QuestionnaireInput {
  answers: QuestionnaireInputAnswers;
  /** @nullable */
  additionalInstructions?: string | null;
}