# lib/api-zod/src/generated/types/questionnaireResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';

export interface QuestionnaireResult {
  needsMoreInfo: boolean;
  followUpQuestions: Question[];
  ready: boolean;
  /** @nullable */
  message?: string | null;
}