# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';
import type { SessionAnswers } from './sessionAnswers';
import type { SessionNotes } from './sessionNotes';
import type { SessionStatus } from './sessionStatus';

export interface Session {
  sessionId: string;
  status: SessionStatus;
  /** @nullable */
  extractedContent?: string | null;
  questions: Question[];
  /** @nullable */
  answers?: SessionAnswers;
  /** @nullable */
  notes?: SessionNotes;
  uploadedImageUrls?: string[];
}