# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';

export interface UploadResult {
  sessionId: string;
  extractedContent: string;
  questions: Question[];
}